ensure open feature supports not only boolean flags
