+++ packages/plugin-e2e/src/options.ts
@@ -10,7 +10,10 @@ export const DEFAULT_ADMIN_USER: User = {
 export const options: Fixtures<{}, PluginOptions> = {
   userPreferences: [{}, { option: true, scope: 'worker' }],
   featureToggles: [{}, { option: true, scope: 'worker' }],
-  openFeatureLatency: [0, { option: true, scope: 'worker' }],
+  openFeature: [
+    { flags: {}, latency: 0 },
+    { option: true, scope: 'worker' },
+  ],
   provisioningRootDir: [path.join(process.cwd(), 'provisioning'), { option: true, scope: 'worker' }],
   user: [DEFAULT_ADMIN_USER, { option: true, scope: 'worker' }],
   grafanaAPICredentials: [DEFAULT_ADMIN_USER, { option: true, scope: 'worker' }],
